Refactor type of field `amount` and `occupied` in `Record`

+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/common/Record.java
@@ -15,9 +15,9 @@ public class Record {
   @SerializedName("ownership")
   public Ownership ownership;
 
-  public long amount;
+  public BigInteger amount;
 
-  public long occupied;
+  public BigInteger occupied;
 
   @SerializedName("asset_info")
   public AssetInfo assetInfo;
